Apply filters to sequence, clean up code
